remove common folder and move components from it to components folder
