+++ packages/frinx-workflow-ui/src/components/workflow-diagram.tsx
@@ -9,9 +9,9 @@ import {
   getElementsFromWorkflow,
   jsonParse,
   unwrap,
-} from '@frinx/shared/src';
+} from '@frinx/shared';
 import { getLayoutedElements } from '../helpers/layout.helpers';
-import { BaseNode, DecisionNode, StartEndNode } from '../components';
+import { BaseNode, DecisionNode, StartEndNode } from './index';
 import { ControlExecutedWorkflowSubscription, Workflow } from '../__generated__/graphql';
 
 const nodeTypes = {
